+++ src/main/java/Data/EkgData.java
@@ -3,8 +3,8 @@ package Data;
 import java.sql.Timestamp;
 
 public interface EkgData {
-    void setVoltage(double voltage);
-    double getVoltage();
+    void setVoltage(int voltage);
+    int getVoltage();
 
     void setTime(Timestamp time);
